requirements.txt
beautifulsoup4==4.10.0
pymongo[srv]==4.0.1
python-telegram-bot==13.11
requests==2.27.1
